goreleaser bug

+++ spectrocloud/utils.go
@@ -27,7 +27,7 @@ func SafeUint32(value int) uint32 {
 	if value < 0 {
 		return 0
 	}
-	if value > math.MaxUint32 {
+	if uint64(value) > math.MaxUint32 {
 		return math.MaxUint32
 	}
 	return uint32(value)
